chore(consent): reorder consent decision DTOs, use cases, and validation schemas
