+++ src/application/index.ts
@@ -3,6 +3,7 @@ export * from './dtos/auth.dtos.js';
 export * from './dtos/validate-client.dtos.js';
 export * from './dtos/user.dtos.js';
 export * from './dtos/client-dtos.js';
+export * from './dtos/consent.dtos.js';
 
 //Use cases
 export * from './use-cases/auth/check-consent.use-case.js';
@@ -11,8 +12,6 @@ export * from './use-cases/auth/generate.auth-code.user-case.js';
 export * from './use-cases/auth/get-jwks.use-case.js';
 export * from './use-cases/auth/login.use-case.js';
 export * from './use-cases/auth/pkce-verifier.use-case.js';
-export * from './use-cases/auth/process-consent.use-case.js';
-export * from './use-cases/auth/show-consent.use-case.js';
 export * from './use-cases/client/create-client.use-case.js';
 export * from './use-cases/client/delete-client.use-case.js';
 export * from './use-cases/client/get-client.use-case.js';
@@ -20,6 +19,8 @@ export * from './use-cases/client/list-client-use-case.js';
 export * from './use-cases/client/rotate-secret.use-case.js';
 export * from './use-cases/client/update-client.use-case.js';
 export * from './use-cases/client/validate-client.use-case.js';
+export * from './use-cases/consent/process-consent.use-case.js';
+export * from './use-cases/consent/show-consent.use-case.js';
 export * from './use-cases/user/get-user.use-case.js';
 export * from './use-cases/user/register-user.use-case.js';
 export * from './use-cases/user/update-password.use-case.js';
@@ -30,3 +31,4 @@ export * from './validate-schemas/auth.schemas.js';
 export * from './validate-schemas/helpers.js';
 export * from './validate-schemas/user.schemas.js';
 export * from './validate-schemas/client.schemas.js';
+export * from './validate-schemas/consent.schema.js';
